# docs/swagger/traPortfolio.v1.yaml
openapi: 3.0.0
info:
  title: traPortfolio
  version: '1.0'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traPortfolio'
  description: traPortfolio v1 API
servers:
  - url: 'http://localhost:1323/api/v1'
    description: local
paths:
...
  /contests:
    get:
      summary: コンテストのリストの取得
      tags:
        - contest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contest'
      operationId: getContests
      description: コンテストのリストを取得します
    parameters: []
    post:
      summary: コンテストの作成
      operationId: createContest
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestDetail'
        '400':
          description: Bad Request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestRequest'
      tags:
        - contest
      description: コンテストを作成します
  '/contests/{contestId}':
    parameters:
      - $ref: '#/components/parameters/contestIdInPath'
    get:
      summary: コンテスト詳細情報の取得
      tags:
        - contest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestDetail'
        '404':
          description: Not Found
      operationId: getContest
      description: コンテスト詳細情報を取得します
    patch:
      summary: コンテスト情報の修正
      operationId: editContest
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテスト情報を修正します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditContestRequest'
      tags:
        - contest
    delete:
      summary: コンテストの削除
      operationId: deleteContest
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストを削除します
      tags:
        - contest
  '/contests/{contestId}/teams':
    parameters:
      - $ref: '#/components/parameters/contestIdInPath'
    get:
      summary: コンテストチームのリストの取得
      tags:
        - contest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContestTeam'
        '404':
          description: Not Found
      operationId: getContestTeams
      description: コンテストのチーム情報を取得します
    post:
      summary: コンテストチームの追加
      operationId: addContestTeam
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestTeam'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストのチーム情報を追加します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContestTeamRequest'
      tags:
        - contest
  '/contests/{contestId}/teams/{teamId}':
    parameters:
      - $ref: '#/components/parameters/contestIdInPath'
      - $ref: '#/components/parameters/teamIdInPath'
    get:
      summary: コンテストチーム詳細情報の取得
      tags:
        - contest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestTeamDetail'
        '404':
          description: Not Found
      operationId: getContestTeam
      description: コンテストチーム詳細情報を取得します
    patch:
      summary: コンテストチームの修正
      operationId: editContestTeam
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストチームを修正します
      tags:
        - contest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditContestTeamRequest'
    delete:
      summary: コンテストチームの削除
      operationId: deleteContestTeam
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストチームを削除します
      tags:
        - contest
  '/contests/{contestId}/teams/{teamId}/members':
    parameters:
      - $ref: '#/components/parameters/contestIdInPath'
      - $ref: '#/components/parameters/teamIdInPath'
    get:
      summary: コンテストチームメンバーの取得
      tags:
        - contest
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
        '404':
          description: Not Found
      operationId: getContestTeamMembers
      description: コンテストチームメンバーを取得します
    post:
      summary: コンテストチームメンバーの追加
      operationId: addContestTeamMembers
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストチームメンバーを追加します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberIDs'
      tags:
        - contest
        - user
    put:
      summary: コンテストチームメンバーの修正
      operationId: editContestTeamMembers
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: コンテストチームメンバーを修正します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberIDs'
      tags:
        - contest
        - user
...
components:
  schemas:
    User:
      title: User
      type: object
      description: ユーザー情報
      properties:
        id:
          type: string
          description: ユーザーUUID
          format: uuid
          x-go-type: uuid.UUID
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]{1,32}$'
          description: ユーザー名
        realName:
          type: string
          description: 本名
      required:
        - id
        - name
        - realName
...
    Contest:
      title: Contest
      type: object
      description: コンテスト情報
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          description: コンテストuuid
        name:
          type: string
          description: コンテスト名
        duration:
          # description: コンテストの開催期間
          $ref: '#/components/schemas/Duration'
      required:
        - id
        - name
        - duration
    ContestDetail:
      title: ContestDetail
      type: object
      description: コンテストの詳細情報
      allOf:
        - $ref: '#/components/schemas/Contest'
        - type: object
          properties:
            link:
              type: string
              format: uri
              description: コンテストの詳細が載っているページへのリンク
            description:
              type: string
              description: コンテストの説明
            teams:
              type: array
              description: コンテストチーム
              items:
                $ref: '#/components/schemas/ContestTeam'
          required:
            - link
            - description
            - teams
    ContestTeam:
      title: ContestTeam
      type: object
      description: コンテストチーム情報
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          description: コンテストチームuuid
        name:
          type: string
          description: チーム名
        result:
          type: string
          description: 順位などの結果
      required:
        - id
        - name
        - result
    ContestTeamDetail:
      title: ContestTeamDetail
      type: object
      description: コンテストチーム詳細情報
      allOf:
        - $ref: '#/components/schemas/ContestTeam'
        - type: object
          properties:
            link:
              type: string
              format: uri
              description: コンテストチームの詳細が載っているページへのリンク
            description:
              type: string
              description: チーム情報
            members:
              type: array
              description: チームメンバーのUUID
              items:
                $ref: '#/components/schemas/User'
          required:
            - link
            - description
            - members
...
    Duration:
      title: Duration
      type: object
      properties:
        since:
          type: string
          format: date-time
          description: 期間始まり
        until:
          type: string
          format: date-time
          description: |-
            期間終わり
            untilがなかったらまだ存続している
      required:
        - since
      description: イベントやコンテストなどの存続期間
...
    MemberIDs:
      title: MemberIDs
      type: object
      description: ユーザーのUUIDの配列
      properties:
        members:
          type: array
          description: ユーザーのUUIDの配列
          items:
            type: string
            format: uuid
            x-go-type: uuid.UUID
      required:
        - members
...
    CreateContestRequest:
      title: CreateContestRequest
      type: object
      description: 新規コンテストリクエスト
      properties:
        name:
          type: string
          description: コンテスト名
        link:
          type: string
          format: uri
          description: コンテストの詳細が載っているページへのリンク
        description:
          type: string
          description: コンテスト説明
        duration:
          # description: コンテストの開催期間
          $ref: '#/components/schemas/Duration'
      required:
        - name
        - description
        - duration
    EditContestRequest:
      title: EditContestRequest
      type: object
      description: コンテスト情報変更リクエスト
      properties:
        name:
          type: string
          description: コンテスト名
        link:
          type: string
          format: uri
          description: コンテストの詳細が載っているページへのリンク
        description:
          type: string
          description: コンテスト説明
        duration:
          # description: コンテストの開催期間
          $ref: '#/components/schemas/Duration'
    AddContestTeamRequest:
      title: AddContestTeamRequest
      type: object
      description: 新規コンテストチームリクエスト
      properties:
        name:
          type: string
          description: チーム名
        link:
          type: string
          format: uri
          description: コンテストチームの説明が載っているページへのリンク
        description:
          type: string
          description: チーム情報
        result:
          type: string
          description: 順位などの結果
      required:
        - name
        - description
    EditContestTeamRequest:
      title: EditContestTeamRequest
      type: object
      description: コンテストチーム情報修正リクエスト
      properties:
        name:
          type: string
          description: チーム名
        link:
          type: string
          format: uri
          description: コンテストチームの説明が載っているページへのリンク
        description:
          type: string
          description: チーム情報
        result:
          type: string
          description: 順位などの結果
  parameters:
...
    contestIdInPath:
      name: contestId
      in: path
      required: true
      description: コンテストUUID
      schema:
        type: string
        format: uuid
        x-go-type: uuid.UUID
    teamIdInPath:
      name: teamId
      in: path
      required: true
      description: コンテストチームUUID
      schema:
        type: string
        format: uuid
        x-go-type: uuid.UUID
...